Separate __test__ directory removed in favor of test files located near system-under-test files
