--- __test__/__mocks__/fileMock.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const rand: number = Math.floor(Math.random() * 6) + 1;
