+++ src/containers/hooks/useWatcher.spec.tsx
@@ -1,13 +1,11 @@
 import React from 'react';
 import { renderHook } from '@testing-library/react-hooks';
-import useWatcher from '../src/containers/hooks/useWatcher';
-import { file, mock, MockInstance } from './__mocks__/mock-fs-promise';
 import { Provider } from 'react-redux';
-import { mockStore } from './__mocks__/reduxStoreMock';
-import { testStore } from './__fixtures__/ReduxStore';
-// import { writeFileAsync } from '../src/containers/io';
 import * as path from 'path';
-// import { rename, unlink } from 'fs-extra';
+import { file, mock, MockInstance } from '../../test-utils/mock-fs';
+import { mockStore } from '../../test-utils/mock-store';
+import { emptyStore } from '../../test-utils/empty-store';
+import useWatcher from './useWatcher';
 
 describe('containers/hooks/useWatcher', () => {
     let mockedInstance: MockInstance;
@@ -21,7 +19,7 @@ describe('containers/hooks/useWatcher', () => {
         });
         return mockedInstance = instance;
     });
-    beforeEach(() => store = mockStore(testStore));
+    beforeEach(() => store = mockStore(emptyStore));
 
     afterAll(() => mockedInstance.reset);
     afterEach(() => {
